# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog Platform Swagger",
    "version": "1.0.0",
    "description": "This is a sample server blog platform Swagger. You can access the Swagger JSON at [http://localhost:5000/swagger.json](http://localhost:5000/swagger.json)"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
    "/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "pincode": {
                    "type": "number"
                  },
                  "phoneNumber": {
                    "type": "number"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "example": {
                  "Status": "Success",
                  "message": "Registration Successfully....",
                  "data": {
                    "name": "mukesh",
                    "email": "mukesh@gmail.com",
                    "address": "55, keval park, punagam",
                    "city": "surat",
                    "pincode": 395010,
                    "phoneNumber": 987653211,
                    "password": 123
                  }
                }
              }
            }
          },
          "409": {
            "description": "User registration failed due to conflict (duplicate name or email)",
            "content": {
              "application/json": {
                "examples": {
                  "user_exists_name": {
                    "value": {
                      "Status": "FAILED",
                      "message": "Name already Exits..!"
                    }
                  },
                  "user_exists_email": {
                    "value": {
                      "Status": "FAILED",
                      "message": "Email already Exits..!"
                    }
                  },
                  "user_exists_both": {
                    "value": {
                      "Status": "FAILED",
                      "message": "Name and Email already Exits..!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "example": {
                  "Status": "Success",
                  "message": "Login Successfully..!",
                  "data": {
                    "_id": 12345,
                    "name": "mukesh",
                    "email": "mukesh@gmail.com",
                    "address": "55, keval park, punagam",
                    "city": "surat",
                    "pincode": 395010,
                    "phoneNumber": 987653211,
                    "password": 123,
                    "tokens": "jwt.token.here"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid password",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Invalid Password"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Invalid Email"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "Status": "ERROR",
                  "message": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User logout",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "Bearer <token>"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "examples": {
                  "user_logout": {
                    "value": {
                      "Status": "Success",
                      "message": "User logout successfully"
                    }
                  },
                  "user_already_logout": {
                    "value": {
                      "Status": "FAILED",
                      "message": "User already logged out"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Unauthorized. User not found."
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "Status": "ERROR",
                  "message": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "/pageinfo": {
      "get": {
        "summary": "Get paginated user list",
        "description": "Retrieve a paginated list of users based on specified filters.",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "name",
            "description": "Filter users by name (case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "email",
            "description": "Filter users by email (case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "address",
            "description": "Filter users by address (case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response containing paginated user list",
            "content": {
              "application/json": {
                "example": {
                  "page": 1,
                  "totalPostes": 1,
                  "totalPages": 1,
                  "postes": {
                    "_id": 12345,
                    "name": "mukesh",
                    "email": "mukesh@gmail.com",
                    "address": "55, keval park, punagam",
                    "city": "surat",
                    "pincode": 395010,
                    "phoneNumber": 987653211,
                    "password": 123
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}